Refactor mcp utils location
