+++ intelli/flow/utils/__init__.py
@@ -6,7 +6,7 @@ This package provides helper utilities for working with Intelli Flows.
 
 # Import utility modules and expose main components
 try:
-    from intelli.utils.mcp_utils import (
+    from intelli.mcp import (
         MCPServerBuilder,
         MCPJSONExtractor,
         create_mcp_preprocessor
